Refactor 'Learn More' form into separate component
