+++ src/components/CourseRecommendationsResults.jsx
@@ -1,48 +1,10 @@
-import { useState } from "react";
 import PropTypes from "prop-types";
 
-import { areInputsValid } from "../utilities";
-
-import ErrorMessage from "./ErrorMessage";
+import LearnMoreForm from "./LearnMoreForm";
 
 import "./CourseRecommendationsResults.css";
 
 function CourseRecommendationsResults({ surveyId, results }) {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [isSubmitConfirmationVisible, setIsSubmitConfirmationVisible] =
-    useState(false);
-
-  const [errors, setErrors] = useState({});
-
-  const getValidationErrors = () => {
-    const newErrors = {};
-
-    if (!name) newErrors.name = "Please enter your name";
-    if (!email) newErrors.email = "Please enter your email";
-
-    return newErrors;
-  };
-
-  const handleNameChange = (event) => {
-    setName(event.target.value);
-  };
-
-  const handleEmailChange = (event) => {
-    setEmail(event.target.value);
-  };
-
-  const submitNameAndEmail = (event) => {
-    event.preventDefault();
-    if (!areInputsValid(getValidationErrors, setErrors)) {
-      setIsSubmitConfirmationVisible(false);
-      return;
-    }
-
-    console.log(name, email);
-    setIsSubmitConfirmationVisible(true);
-  };
-
   return (
     <div className="course-recommendations-results">
       <h4>Recommended Courses</h4>
@@ -92,42 +54,7 @@ function CourseRecommendationsResults({ surveyId, results }) {
         custom designed program to your employees, input your name and email
         below and a member from our team will contact you in 24 hours.
       </p>
-
-      <form
-        className="submit-name-and-email-form"
-        onSubmit={submitNameAndEmail}
-      >
-        <div className="submit-name-and-email-container">
-          <div className="form-element">
-            <label htmlFor={`name-${surveyId}`}>Name</label>
-            <input
-              id="name"
-              name="name"
-              value={name}
-              onChange={handleNameChange}
-            />
-            <ErrorMessage message={errors.name} />
-          </div>
-
-          <div className="form-element">
-            <label htmlFor={`email-${surveyId}`}>Email</label>
-            <input
-              id="email"
-              name="email"
-              value={email}
-              onChange={handleEmailChange}
-            />
-            <ErrorMessage message={errors.email} />
-          </div>
-        </div>
-
-        <button type="submit">Submit</button>
-      </form>
-      {isSubmitConfirmationVisible && (
-        <p className="submit-confirmation">
-          Submitted! We will be in touch soon.
-        </p>
-      )}
+      <LearnMoreForm surveyId={surveyId} />
     </div>
   );
 }
