+++ src/components/LearnMoreForm.css
@@ -1,0 +1,15 @@
+.submit-name-and-email-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 0 calc(min(200px, 7vw) + min(250px, 10vw));
+  gap: 20px;
+}
+
+.submit-name-and-email-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  gap: 10px;
+}
