+++ src/components/ROICalculationResults.css
@@ -9,9 +9,3 @@
   flex-direction: column;
   align-items: center;
 }
-
-.submit-confirmation {
-  font-size: 0.75em;
-  font-weight: bold;
-  text-align: center;
-}
